Removed Externs.Impl. Since we got this working on >NETCORE3_1 with the same code for NETCORE3_1, we're not doing to need to conditionalize on higher Core versions.
