+++ src/IKVM.Runtime/Java/Externs/sun/security/ec/ECDSASignature.cs
@@ -1,4 +1,14 @@
-﻿namespace IKVM.Runtime.Java.Externs.sun.security.ec
+﻿using System;
+using System.Security.Cryptography;
+
+#if FIRST_PASS == false
+
+using java.lang;
+using java.security;
+
+#endif
+
+namespace IKVM.Java.Externs.sun.security.ec
 {
 
     /// <summary>
@@ -20,7 +30,30 @@
 #if FIRST_PASS
             throw new NotImplementedException();
 #else
-            return IKVM.Java.Externs.Impl.sun.security.ec.ECDSASignature.signDigest(digest, s, encodedParams, seed);
+#if NET47_OR_GREATER || NETCOREAPP3_1_OR_GREATER
+            var curve = ECUtil.DecodeParameters(encodedParams);
+            if (curve.IsNamed == false)
+                throw new InvalidAlgorithmParameterException();
+
+            try
+            {
+                // import key as private key
+                var p = new ECParameters();
+                p.Curve = curve;
+                p.D = s;
+                using var dsa = ECDsa.Create(p);
+
+                // sign the digest
+                var r = dsa.SignHash(digest);
+                return r;
+            }
+            catch (CryptographicException e)
+            {
+                throw new IllegalStateException(e);
+            }
+#else
+            throw new NotImplementedException();
+#endif
 #endif
         }
 
@@ -37,7 +70,30 @@
 #if FIRST_PASS
             throw new NotImplementedException();
 #else
-            return IKVM.Java.Externs.Impl.sun.security.ec.ECDSASignature.verifySignedDigest(signature, digest, w, encodedParams);
+#if NET47_OR_GREATER || NETCOREAPP3_1_OR_GREATER
+            var curve = ECUtil.DecodeParameters(encodedParams);
+            if (curve.IsNamed == false)
+                throw new InvalidAlgorithmParameterException();
+
+            try
+            {
+                // import key as private key
+                var p = new ECParameters();
+                p.Curve = curve;
+                p.Q = ECUtil.ImportECPoint(w);
+                using var dsa = ECDsa.Create(p);
+
+                // verify the signature
+                var r = dsa.VerifyHash(digest, signature);
+                return r;
+            }
+            catch (CryptographicException e)
+            {
+                throw new IllegalStateException(e);
+            }
+#else
+            throw new NotImplementedException();
+#endif
 #endif
         }
