+++ src/IKVM.Runtime/Java/Externs/sun/security/ec/ECPoint.cs
@@ -1,4 +1,4 @@
-﻿namespace IKVM.Java.Externs.Impl.sun.security.ec
+﻿namespace IKVM.Java.Externs.sun.security.ec
 {
 
 #if NET47_OR_GREATER || NETCOREAPP
